24_Read処理のServiceとController部分を実装　課題
