+++ src/main/java/shohei/student/management/repository/StudentRepository.java
@@ -6,8 +6,20 @@ import org.apache.ibatis.annotations.Select;
 import shohei.student.management.data.Courses;
 import shohei.student.management.data.Student;
 
+/**
+ * 受講生情報を扱うリポジトリ。
+ * <p>
+ * 全件検索や単一条件での検索、コース情報の検索が行えるクラスです。
+ */
+
 @Mapper
 public interface StudentRepository {
+
+  /**
+   * 全件検索をします。
+   *
+   * @return 全件検索した受講生情報の一覧。
+   */
 
   @Select("SELECT * FROM students")
   List<Student> search();
